Completed basic api

+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Entities/Article.cs
@@ -6,9 +6,8 @@ public class Article : BaseEntity
     public string Title { get; set; }
     public string Content { get; set; }
     public string Abstraction { get; set; }
-    public long UserId { get; set; }
-    public long TopicId { get; set; }
-    //public string Tags { get; set; }
+    public int UserId { get; set; }
+    public int TopicId { get; set; }
     public int SortNumber { get; set; }
     public int IsHidden { get; set; }
 }
