+++ BackendDotNet/FF.Articles.Backend/FF.Articles.Backend.Contents.API/Models/Entities/ArticleTag.cs
@@ -1,8 +1,11 @@
 ﻿using FF.Articles.Backend.Common.Bases;
 
 namespace FF.Articles.Backend.Contents.API.Models.Entities;
+/// <summary>
+/// Ignore BaseEntity optional columns
+/// </summary>
 public class ArticleTag : BaseEntity
 {
-    public long ArticleId { get; set; }
-    public long TagId { get; set; }
+    public int ArticleId { get; set; }
+    public int TagId { get; set; }
 }
